Move files to match established structure
